Fixes from 1st part

+++ AVT_start/AVT_start/Engine/HeaderFiles/mxfactory.h
@@ -13,17 +13,17 @@ public:
 
 	// Matrix3: identity, dual
 	Matrix3 identity3();
-	Matrix3 dual(float ax, float ay, float az);
-	Matrix3 squaredual(float ax, float ay, float az);
+	Matrix3 dual(double ax, double ay, double az);
+	Matrix3 squaredual(double ax, double ay, double az);
 
 	// Matrix4: identity, scaling, translation, rotation (Rodrigues)
 	Matrix4 identity4();
-	Matrix4 scaling4(float sx, float sy, float sz);
-	Matrix4 translation4(float tx, float ty, float tz);
-	Matrix4 rotation4(float rx, float ry, float rz, float theta);
-	Matrix4 invscaling4(float sx, float sy, float sz);
-	Matrix4 invtranslation4(float tx, float ty, float tz);
-	Matrix4 invrotation4(float rx, float ry, float rz, float theta);
+	Matrix4 scaling4(double sx, double sy, double sz);
+	Matrix4 translation4(double tx, double ty, double tz);
+	Matrix4 rotation4(double rx, double ry, double rz, double theta);
+	Matrix4 invscaling4(double sx, double sy, double sz);
+	Matrix4 invtranslation4(double tx, double ty, double tz);
+	Matrix4 invrotation4(double rx, double ry, double rz, double theta);
 
 };
